requirements.txt
# requirements.txt
# List required Python packages here 
streamlit
pandas
dnspython